chore: Move remaining functions over
