+++ mobile/src/services/RNTPService.ts
@@ -6,8 +6,8 @@ import TrackPlayer, {
 } from "@weights-ai/react-native-track-player";
 
 import i18next from "~/modules/i18n";
-import { addPlayedTrack } from "~/api/recent";
 import { deleteTracks } from "~/api/track";
+import { addPlayedTrack } from "~/data/recent/api";
 import { playbackStore } from "~/stores/Playback/store";
 import { PlaybackControls, Queue } from "~/stores/Playback/actions";
 import { preferenceStore } from "~/stores/Preference/store";
